mem: move mempool util from ngs
