+++ vita3k/ngs/CMakeLists.txt
@@ -6,7 +6,6 @@ add_library(
 	include/ngs/modules/player.h
     include/ngs/modules/passthrough.h
 	include/ngs/common.h
-	include/ngs/mempool.h
 	include/ngs/scheduler.h
 	include/ngs/state.h
 	include/ngs/system.h
@@ -15,7 +14,6 @@ add_library(
 	src/modules/master.cpp
 	src/modules/player.cpp
     src/modules/passthrough.cpp
-	src/allocator.cpp
 	src/ngs.cpp
 	src/route.cpp
 	src/scheduler.cpp)
